python/console/requirements.txt
python-dotenv
openai
openai-agents
pydantic